Se arreglan los tipos de datos de la API

+++ app/src/main/java/ar/edu/unlam/mobile/scaffold/ui/screens/transactionScreen/TransactionScreen.kt
@@ -148,7 +148,7 @@ fun TransactionScreen(
             viewModel.getCurrencyConversion(
                 source = "USD",
                 target = "ARS",
-                quantity = amount.toDouble(),
+                quantity = amount,
             )
         }) {
             Text(text = "Convertir")
